requirements.txt
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.1
matplotlib==3.7.1
seaborn==0.12.2
scipy==1.10.0
Flask==3.0.3
joblib==1.4.2
google-cloud-storage==2.18.2
pylint==2.4.4
pytest==6.2.0